feat: evaluate expression

+++ migrations/000001_init_database.up.sql
@@ -2,7 +2,7 @@ CREATE TYPE task_status AS ENUM ('SCHEDULED', 'COMPLETED', 'FAILED');
 
 CREATE TABLE tasks (
   id SERIAL PRIMARY KEY,
-  status task_status ,
+  status task_status,
   expression VARCHAR(150),
-  result BIGINT
+  result NUMERIC
 );
